Port integration config views to TypeScript/spina.
This change converts the integration config list view and pop-up to
TypeScript and spina.

This also makes a few small updates to our packaging for djblets to
include jasmine-suites as a dependency. This was used in djblets but not
included in any of our build stuff. It didn't actually matter for
runtime use when running the tests via our RB js-tests endpoint, but
it's necessary for running the TS compiler for files that import
jasmine within the djblets directory.

Testing Done:
- Ran js-tests.
- Added, edited, and deleted integration configs.

Reviewed at https://reviews.reviewboard.org/r/14103/