Corrected, additional fixes for configforms labels.
In a previous change, I added `:not(.ink-c-button__label)` to the
`label` rule for configforms. This prevented button labels from having
odd styling, but had the side effect of bumping up the precedence of
that rule, which then caused it to take priority over the label rules
for action drop-downs.

The entire configforms framework feels like it's in dire need of a
rewrite, but for the moment, it turns out that the only <label> in any
of our configforms that wasn't already using other rules
(`.rb-c-form-field__label`) was the avatar backend. This was because we
had an template override for this form which was adding the label and
form field directly, instead of just including the standard configforms
field template.

This change makes several small fixes to the label situation:
- Remove the display/margin rules for generic `<label>` tags.
- Changed the avatar settings form template to use the configforms
  template for the avatar backend field.
- Fixed that template to not overwrite the `form` variable in a loop.
- Added `<input type="tel">` to our standard input style rules.

Testing Done:
- Went through all of the "My Account" and admin configforms and
  verified that labels on form fields looked correct.
- Made sure the radio button drop-downs in the API tokens list worked
  correctly.

Reviewed at https://reviews.reviewboard.org/r/14079/