Add dark mode styles for related object selector and fix table mode.
This change adds dark mode styling (and generally overrides the default
Selectize view) for the related object selector. This isn't something
that's currently visible in Review Board, since the admin interface is
always in light mode, but it helps in RBCommons or other projects using
Djblets.

Also, I hadn't noticed that the option rendering change to allow
subclasses to use a table for the selected options broke the selectize
drop-down. I've changed it so there are now separate functions for the
drop-down and selected options (the default implementation of the former
just calling the former), so that we can provide a separate layout for
the drop-down and the selected items.

Testing Done:
- Verified that both the drop-down and selected items lists looked
  correct for both <table> and <li> based selectors.
- Verified that the selectors looked good in both light and dark modes.

Reviewed at https://reviews.reviewboard.org/r/14148/