Add several build files to the Djblets source distribution.
There are several files we need these days for building and testing the
Djblets packages which weren't included in the source distribution. This
includes the static media toolchain, the unit test configuration, and
the tox file.

This change adds all these to the manifest, ensuring they're there in
the tarball. It's a precursor to further work on the Djblets packaging.

Testing Done:
Built a source distribution and verified the package contents.

Tested this along with in-progress package management code, and managed
to build wheels from the source distribution that were identical to
wheels built from the tree.

Reviewed at https://reviews.reviewboard.org/r/14136/