Fix a regression where selected related objects wouldn't render properly.
Our base `RelatedObjectSelectorView` had a regression where selected
objects would not longer render properly. This happened because we
introduced a new `renderSelectedOption()` which defaults to calling
`renderOption()`, but we forgot to actually pass the item to
`renderOption()` here. This change fixes that.

This change also updates the remove icon for objects in the related
object selector to be our standard Ink delete icon. And we enforce the
hover/not-hovered colors for the icon, which fixes a bug with Review
Board's related user selector where the icon's hover color was blue due
to being a link element.

Testing Done:
- Ran JS unit tests.
- Tested using the user, group and repository selector widgets.

Reviewed at https://reviews.reviewboard.org/r/14329/