Add typing for condition choices.
This introduces typing for `BaseConditionChoice` and all subclasses.
All class attributes are clearly typed, as are instance variables.

This is a pretty straight-forward change. However, there are some issues
with the current design that prevents us from correctly typing the
`default_value_field`. In some cases, we set this as an attribute, and
in others as a function. That will need to be cleaned up as a separate
change.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14266/