Support ConditionChoices instances in ConditionsField and add typing.
`ConditionsField` claimed support for taking a `ConditionChoices` class
or a callable returning the class, but not an instance. We have call
sites that passed in a class accordingly. The problem with prioritizing
classes is that there's no ability for code to register choices into a
`ConditionChoices` after construction (say, for extensions).

In practice, this field supported instances just fine, and we even had
unit tests for it, but we now document this explicitly (fixing class
names in the process).

The field has also been updated with typing, signature matches, and
some code cleanup.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14272/