Add log timers and better logging around DMARC record fetching.
In the event of a DNS issue, it's helpful to know how long a process
took to attempt the resolution, and to know what the failure was. To
help with this, we now wrap the DNS fetching with `log_timed()` and have
logging for different failure types.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14366/