Fix a regression in assertQueries and update our dependency.
`assertQueries()` was updated to wrap `django-assert-queries`, but this
ended up being invoked incorrectly, causing the assertion to always
succeed. We were yielding the result and not entering its context.

This change updates the call to simply return the `assert_queries()`'s
context manager directly. This avoids the issue and avoids wrapping in
another layer of context managers.

We also now depend on django-assert-queries 2.0.1, which, along with the
above, has the added benefit of not including `assertQueries()` in the
stack trace.

Testing Done:
Djblets unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/14384/