Enhance typing for KeyValueForm.
This change makes it so KeyValueForm is typed as a generic, allowing us
to specify the type of the `instance` member which is operated on. The
default implementations of get and set assume a dict-like interface, so
I've additionally added some runtime-checkable protocols for those.

Testing Done:
- Ran unit tests.
- Verified that type checkers were happy when creating KeyValueForm
  subclasses.

Reviewed at https://reviews.reviewboard.org/r/14437/