Add extra_context support, typing, and docs for extension views.
The extension views for listing extensions and configuring an extension
were far too basic and limiting. There was no way of getting extra
context into the template without storing it either on the form or the
extension.

This change adds an `extra_context` parameter to both the
`extension_list` and `configure_extension` views for providing this
information.

It also adds documentation (which was missing completely) and type
annotations to make it easier to know what's expected to be passed to
these views. The arguments are now expected to be passed as keyword
arguments as well, which is the default case when included directly in
URL patterns.

Testing Done:
Unit tests pass.

Tested this with an extension.

Reviewed at https://reviews.reviewboard.org/r/14500/