Allow DJBLETS_HEALTHCHECK_IPS to list subnets.
In some larger setups, the individual IP of the management node for a
cluster may not be fixed. This makes it very annoying to set the allowed
IP whitelist for the healthcheck endpoint.

This change makes it so the `DJBLETS_HEALTHCHECK_IPS` setting is now
formally a comma-separated list which can include either individual IP
addresses or CIDR subnets.

This change also adds unit tests for the healthcheck endpoint, which
didn't have any.

Testing Done:
- Verified that both individual IP and subnet matching worked as
  expected.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14579/