Fix name of `find_all` parameter in staticfiles finder.
Our staticfiles finder for extensions was using the name `all` for the
optional second parameter, but inside Django this has been called
`find_all`. This value is now getting passed in as a kwarg, so we need
the name of that parameter to match.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14768/