Add @rollup/plugin-commonjs as a devDependency.
The new rollup config in Djblets uses the commonjs plugin, but this was
only installed as part of the Review Board dependencies. When doing a
full environment install with both Djblets and Review Board, everything
worked fine, but when doing individual builds of Djblets in its own
environment, this would fail.

Testing Done:
Built a wheel of Djblets in isolation.

Reviewed at https://reviews.reviewboard.org/r/14753/